add shared module for material module exports
